🏗️ Lastly exploded the main measurement file.
